SIMSBIOHUB-61: Survey publish modals refactor WIP
